Add event-driven order status notifications, refactor Notification
Implemented event-driven notification system for order status updates using an in-memory event bus and handler. Refactored Notification.EntityId from long to int, including migration and model updates. Updated dependency injection and controller logic to support new architecture. Added supporting interfaces and classes for event handling.

+++ Models/Notification.cs
@@ -25,7 +25,7 @@ namespace Models
 
         // Polymorphic reference → for deep linking with enities
         public string? EntityType { get; set; }   // Order, Product, Offer
-        public long? EntityId { get; set; }
+        public int? EntityId { get; set; }
 
         public bool IsRead { get; set; } = false;
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
